EVIO fixes

+++ src/plugins/monitoring/ST_online_Tresolution/JEventProcessor_ST_online_Tresolution.cc
@@ -70,7 +70,7 @@ jerror_t JEventProcessor_ST_online_Tresolution::init(void)
 //------------------
 // brun
 //------------------
-jerror_t JEventProcessor_ST_online_Tresolution::brun(JEventLoop *eventLoop, int runnumber)
+jerror_t JEventProcessor_ST_online_Tresolution::brun(JEventLoop *eventLoop, int32_t runnumber)
 {
 	// This is called whenever the run number changes
   // Get the particleID object for each run
@@ -120,7 +120,7 @@ jerror_t JEventProcessor_ST_online_Tresolution::brun(JEventLoop *eventLoop, int 
 //------------------
 // evnt
 //------------------
-jerror_t JEventProcessor_ST_online_Tresolution::evnt(JEventLoop *loop, int eventnumber)
+jerror_t JEventProcessor_ST_online_Tresolution::evnt(JEventLoop *loop, uint64_t eventnumber)
 {
 	// This is called for every event. Use of common resources like writing
 	// to a file or filling a histogram should be mutex protected. Using
